requirements.txt
kivy==2.3.0
kivymd==1.2.0
pandas==2.1.4
numpy==1.24.3
matplotlib==3.8.2
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
google-generativeai==0.3.2
Pillow==10.1.0
chardet==5.2.0
supabase==2.25.0
httpx==0.25.2
postgrest==1.3.0
realtime==1.1.1
gotrue==1.4.0
storage3==0.9.0